機能追加、修正(#1, #2)

+++ FistirDictionary/FDictionary.cs
@@ -14,6 +14,7 @@ using Microsoft.EntityFrameworkCore.Sqlite;
 using Microsoft.EntityFrameworkCore.Storage;
 using System.Text.RegularExpressions;
 using DocumentFormat.OpenXml.Packaging;
+using System.IO;
 
 namespace FistirDictionary
 {
@@ -47,6 +48,33 @@ namespace FistirDictionary
     /// </summary>
     internal class FDictionary
     {
+        public class DictionaryNotFoundExcepction : Exception
+        {
+            public string? Message { get; set; }
+            public DictionaryNotFoundExcepction(string message) : base(message)
+            {
+                Message = message;
+            }
+        }
+
+        public class DictionaryException : Exception
+        {
+            public string? Message { get; set; }
+            public DictionaryException(string message) : base(message)
+            {
+                Message = message;
+            }
+        }
+
+        public class DictionaryBrokenException : Exception
+        {
+            public string? Message { get; set; }
+            public DictionaryBrokenException(string message) : base(message)
+            {
+                Message = message;
+            }
+        }
+
         private class WordRhyme
         {
             public int WordID { get; set; }
@@ -66,11 +94,15 @@ namespace FistirDictionary
 
         public static ICollection<Word> GetDictionaryMetadata(string dictionaryPath)
         {
+            if (!System.IO.File.Exists(dictionaryPath))
+            {
+                throw new DictionaryNotFoundExcepction($"{dictionaryPath} が見つかりません。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
             return (from word in db.Words
-                   where word.Headword != null && word.Headword.StartsWith("__")
-                   select word)
-                   .ToArray();
+                where word.Headword != null && word.Headword.StartsWith("__")
+                select word)
+                .ToArray();
         }
 
         /// <summary>
@@ -92,6 +124,10 @@ namespace FistirDictionary
             bool enableWordHistory,
             string authorInfo)
         {
+            if (System.IO.File.Exists(dictionaryPath))
+            {
+                throw new DictionaryException($"{dictionaryPath} がすでに存在します。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
             RelationalDatabaseCreator rdc = (RelationalDatabaseCreator)db.Database.GetService<IDatabaseCreator>();
             if (!rdc.HasTables())
@@ -102,8 +138,6 @@ namespace FistirDictionary
                 {
                     { "__Name", dicname },
                     { "__Description", description },
-                    { "__ScansionScript", scansionScript },
-                    { "__DerivationScript", derivationScript },
                     { "__EnableHistory", enableWordHistory ? "true" : "false" },
                     { "__Author", authorInfo }
                 };
@@ -129,6 +163,10 @@ namespace FistirDictionary
             string translation,
             string example)
         {
+            if (!System.IO.File.Exists(dictionaryPath))
+            {
+                throw new DictionaryNotFoundExcepction($"{dictionaryPath} が見つかりません。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
             var w = new Word
             {
@@ -144,6 +182,10 @@ namespace FistirDictionary
 
         public static int AddWord(string dictionaryPath, Word[] wordsToAdd)
         {
+            if (!System.IO.File.Exists(dictionaryPath))
+            {
+                throw new DictionaryNotFoundExcepction($"{dictionaryPath} が見つかりません。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
             db.Words.AddRange(wordsToAdd);
             return db.SaveChanges(true);
@@ -156,6 +198,10 @@ namespace FistirDictionary
             string translation,
             string example)
         {
+            if (!System.IO.File.Exists(dictionaryPath))
+            {
+                throw new DictionaryNotFoundExcepction($"{dictionaryPath} が見つかりません。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
             var w = (from word in db.Words
                      where word.WordID == wordID
@@ -207,6 +253,10 @@ namespace FistirDictionary
 
         public static Word GetWord(string dictionaryPath, int wordID)
         {
+            if (!System.IO.File.Exists(dictionaryPath))
+            {
+                throw new DictionaryNotFoundExcepction($"{dictionaryPath} が見つかりません。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
             return (from word in db.Words
                     where word.WordID == wordID
@@ -216,21 +266,36 @@ namespace FistirDictionary
 
         public static void DeleteWord(string dictionaryPath, int wordID)
         {
+            if (!System.IO.File.Exists(dictionaryPath))
+            {
+                throw new DictionaryNotFoundExcepction($"{dictionaryPath} が見つかりません。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
             var w = db.Words.Single(word => word.WordID == wordID);
             db.Words.Remove(w);
             db.SaveChanges();
         }
 
-        public static Word[] SearchWord(string dictionaryPath, SearchStatement[] statements, bool ignoreCase)
+        public static Word[] SearchWord(string dictionaryPath, SearchStatement[] statements, bool ignoreCase, string scansionScriptPath)
         {
+            if (!File.Exists(dictionaryPath))
+            {
+                throw new DictionaryNotFoundExcepction($"{dictionaryPath} が見つかりません。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
-            var scansionScriptPath = db.Words.First(word => word.Headword == "__ScansionScript").Translation;
             var dictionaryName = db.Words.First(word => word.Headword == "__Name").Translation;
             string scriptHash = "";
-            if (scansionScriptPath != null && scansionScriptPath.Length > 0)
+            var rhymeRequired = statements.Where(statement => statement.Target == SearchTarget.Rhyme).Count() > 0;
+            if (rhymeRequired && scansionScriptPath != null && scansionScriptPath.Length > 0)
             {
-                scriptHash = FLua.GetScriptHash(scansionScriptPath);
+                if (File.Exists(scansionScriptPath))
+                {
+                    scriptHash = FLua.GetScriptHash(scansionScriptPath);
+                }
+                else
+                {
+                    throw new FileNotFoundException($"{scansionScriptPath}が見つかりません。");
+                }
             }
             IQueryable<WordRhyme> dbQuery =
                 from word in db.Words
@@ -432,7 +497,6 @@ namespace FistirDictionary
                             case SearchTarget.Headword:
                                 if (ignoreCase)
                                 {
-
                                     dbQuery = dbQuery.Where(word =>
                                         word.Headword != null && EF.Functions.Like(word.Headword.ToLower(), $"%{statement.Keyword.ToLower()}"));
                                 }
@@ -604,6 +668,10 @@ namespace FistirDictionary
 
         public static List<WordHistory> GetWordHistory(string dictionaryPath, int wordID)
         {
+            if (!System.IO.File.Exists(dictionaryPath))
+            {
+                throw new DictionaryNotFoundExcepction($"{dictionaryPath} が見つかりません。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
             var result = db.WordHistories.Where(wh => wh.WordID == wordID).ToList();
             return result;
@@ -611,6 +679,10 @@ namespace FistirDictionary
 
         public static int GetWordCount(string dictionaryPath)
         {
+            if (!System.IO.File.Exists(dictionaryPath))
+            {
+                throw new DictionaryNotFoundExcepction($"{dictionaryPath} が見つかりません。");
+            }
             using var db = new DictionaryContext(GetSqliteConnectionString(dictionaryPath));
             var result = db.Words.Where(word => word.Headword != null && !word.Headword.StartsWith("__")).Count();
             return result;
